Align color swaths in columns instead of rows
Problem:

Base16 color schemes are usually composed of eight greyscale colors
and eight more vibrant colors.
Typically these correspond to backgrounds and foregrounds, although that
is not always the case.

When we lay out colors in rows,
it is difficult to tell how the background and foreground colors
are related.

Solution:

Switch the colors to be laid out based on columns instead of rows,
so the background colors take up the left column,
and the foreground colors take up the right.