Set up palette with hue and lightness varying
We chose the Hue/Saturation/Lightness color scale
to display our palette
because it is one of the easiest for humans to navigate visually.

The HTML canvas element only accepts color data
in the RGB (Red/Green/Blue) color space,
so we need to convert from HSL to RGB
for each color we want to display.

The HSL-to-RGB transformation algorithm
was taken from http://stackoverflow.com/a/9493060/1857539