Change color scheme when user selects from palette
Feature:

When a user clicks on the color palette,
the corresponding color square and preview text
should change to match the selected color,
so the user can modify the color scheme to their liking.

Limitations:

This commit only allows users to edit the "base08" color,
which is used for variable names in the preview.

Future commits should allow users to select
which color they want to change.