Outline selected color square
Problem:

When the user selects a color square to change it,
there is no indication that the square is selected.
This makes it difficult for the user
to tell if anything happened when they clicked,
and there is no reminder about what color they are changing.

Solution:

Add a black outline around the selected color square
to remind the user what color they are editing.