Allow user to change background color
Problem:

The background color for the color scheme preview
is hard-coded to the initial value of the `base00` color.
It's impossible for the user to change the background color.

Solution:

Refactor the CSS to add `fg-` and `bg-` classes for each of the colors.
The javascript can target these prefixes
to change the foreground or background color of elements,
as appropriate.