Live-update color as user drags mouse on palette
Problem:

In order to browse for colors,
or drill down on a specific color,
the user must repeatedly click on the palette
to see the color swatch and preview update.

Solution:

Fire the color-change event
when the user drags their mouse on the palette.
This allows the user to see the color swatch and preview
update instantaneously as they move their cursor across the palette.