Load colors from URL hash
Feature:

When I open the URL for a color scheme I like,
I want to see the colors loaded up immediately on the page
so I can preview and download the color scheme for myself.

Implementation:

Use the `window.location.hash` API
to fetch the color scheme from the URL.

Minor Changes:

Update `ColorHex` to accept colors with or without a leading `#`.