Refactor: Add `Swatch#getHexString()`
Problem:

`application.js` does a lot of work to convert the swatch's colors
into a format acceptable for the page URL.

In the future, we would need to repeat that work
in order to send the current color scheme to the server for downloads.

Solution:

Move that work into the `Swatch` class,
to DRY up the code.