Bugfix: pad hex values with leading zeros
Problem:

When converting from RGB to hex,
when one of the color values was less than 16,
the converted hex value for that channel
does not include a leading zero.

This results in an invalid hex value,
which means the color doesn't show up correctly
in the swatch or preview.

Solution:

Correctly pad the hex values when converting from RGB.