Use `sprockets-es6` gem to compile javascripts
Problem:

Although the application works locally,
it was unable to deploy to Heroku
because of Javascript compilation errors.

This is likely because we are using es6 features
that are supported by Chrome
but not supported by our Javascript runtime, ExecJS.

Solution:

Include `sprockets-es6` for compiling es6 syntax,
and add an `.es6` extension to all files that use es6 syntax.