Refactor config generation to be more streamlined
Problem:

In order to generate configs,
we had to execute the base16-builder script
and manually manipulate the generated files into a ZIP file.

This process was brittle.
It overwrote files at the same location for each request,
which made it vulnerable to race conditions.
In addition, Heroku doesn't play nicely with the filesystem,
so the system flat-out didn't work on Heroku.

Solution:

- Pass config directly into base16 theme builder
  instead of writing a yaml scheme file
- Create & send zip file with IO streams, not files
- delete templates that throw errors