Listen to user scroll events to change saturation
Feature:

When I select a color on the palette,
I want to easily change the saturation
so I can access a wider variety of colors.

Solution:

Tie vertical scrolling on the color palette
to the saturation of the colors on the palette.
As the user scrolls, the color space will change accordingly.

We clamp saturation between 0 and 1, because that's the standard range
that the HSL color space works with.

We chose a scrolling-to-saturation ratio of 1/500,
because it strikes a good balance between responsiveness
and fine-grained control.